added manager 2.0

+++ StoreHelper.BLL.Contracts/IPurchaseManager.cs
@@ -10,6 +10,6 @@ namespace StoreHelper.BLL.Contracts
 {
     public interface IPurchaseManager
     {
-        ProductDto MakePurchase(List<int> productIds);
+        ProductDto MakePurchase(List<long> productIds);
     }
 }
