+++ StoreHelperDAL/Models/Product.cs
@@ -8,12 +8,12 @@ using StoreHelper.Dal.Core.Interfaces.Entity;
 
 namespace StoreHelperDAL.Models
 {
-    public class Product : IKeyable<int>
+    public class Product : IKeyable<long>
     {
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        public  int Id { get; set; }
-        public  string Name { get; set; }
-        public  double Price { get; set; }
+        public long Id { get; set; }
+        public string Name { get; set; }
+        public double Price { get; set; }
 
         public ProductType ProductType { get; set; }
         public virtual ICollection<Purchase> Purchases { get; set; }
